- freezing py3 branch before opening a new one for jit compiling  On branch py3  Your branch is up to date with 'origin/py3'.
 Changes to be committed:
	modified:   pyLpov/proc/processing.py
	modified:   pyLpov/scripts/scenarios/ssvep_online.py

+++ pyLpov/proc/processing.py
@@ -11,7 +11,8 @@ import numpy as np
 # Filter
 def eeg_filter(eeg, fs, low_pass, high_pass, order):
     # assert( high_pass > low_pass), 'Band-pass filtering: High cutoff frequency should be higher than low cutoff frequency'
-    B, A = butter(order, np.array([low_pass, high_pass], dtype=float)/(fs/2), btype='bandpass')
+    # B, A = butter(order, np.array([low_pass, high_pass], dtype=float)/(fs/2), btype='bandpass')
+    B, A = butter(order, np.array([low_pass, high_pass])/(fs/2), btype='bandpass')
     return filtfilt(B, A, eeg, axis=0)
     # return sig.filtfilt(B, A, eeg, axis=0, padtype='odd', padlen=3*(max(len(B),len(A))-1))
 '''
